Refactored Stats Iterator code to KryoLazyStatsUtils
