+++ geomesa-accumulo/geomesa-accumulo-datastore/src/main/scala/org/locationtech/geomesa/accumulo/data/stats/AccumuloGeoMesaStats.scala
@@ -20,11 +20,11 @@ import org.geotools.data.{Query, Transaction}
 import org.joda.time._
 import org.locationtech.geomesa.accumulo.AccumuloVersion
 import org.locationtech.geomesa.accumulo.data.{AccumuloBackedMetadata, _}
-import org.locationtech.geomesa.accumulo.iterators.KryoLazyStatsIterator
 import org.locationtech.geomesa.filter._
 import org.locationtech.geomesa.index.conf.QueryHints
 import org.locationtech.geomesa.index.stats.MetadataBackedStats.KeyAndStat
 import org.locationtech.geomesa.index.stats._
+import org.locationtech.geomesa.index.utils.KryoLazyStatsUtils
 import org.locationtech.geomesa.utils.collection.SelfClosingIterator
 import org.locationtech.geomesa.utils.geotools.RichSimpleFeatureType.RichSimpleFeatureType
 import org.locationtech.geomesa.utils.geotools.SimpleFeatureTypes
@@ -97,7 +97,7 @@ class AccumuloGeoMesaStats(val ds: AccumuloDataStore, statsTable: String, val ge
       val reader = ds.getFeatureReader(query, Transaction.AUTO_COMMIT)
       val result = try {
         // stats should always return exactly one result, even if there are no features in the table
-        KryoLazyStatsIterator.decodeStat(reader.next.getAttribute(0).asInstanceOf[String], sft)
+        KryoLazyStatsUtils.decodeStat(reader.next.getAttribute(0).asInstanceOf[String], sft)
       } finally {
         reader.close()
       }
