+++ geomesa-accumulo/geomesa-accumulo-datastore/src/main/scala/org/locationtech/geomesa/accumulo/index/legacy/attribute/AttributeQueryableIndex.scala
@@ -23,7 +23,7 @@ import org.locationtech.geomesa.features.SerializationType
 import org.locationtech.geomesa.filter._
 import org.locationtech.geomesa.filter.visitor.FilterExtractingVisitor
 import org.locationtech.geomesa.index.api.FilterStrategy
-import org.locationtech.geomesa.index.utils.Explainer
+import org.locationtech.geomesa.index.utils.{Explainer, KryoLazyStatsUtils}
 import org.locationtech.geomesa.utils.geotools.RichAttributeDescriptors.RichAttributeDescriptor
 import org.locationtech.geomesa.utils.geotools.RichSimpleFeatureType.RichSimpleFeatureType
 import org.locationtech.geomesa.utils.index.{IndexMode, VisibilityLevel}
@@ -134,7 +134,7 @@ trait AttributeQueryableIndex extends AccumuloFeatureIndex with LazyLogging {
       if (descriptor.getIndexCoverage() == IndexCoverage.FULL) {
         val iter = KryoLazyStatsIterator.configure(sft, this, filter.secondary, hints, hasDupes)
         val iters = visibilityIter(sft) :+ iter
-        val reduce = Some(KryoLazyStatsIterator.reduceFeatures(sft, hints)(_))
+        val reduce = Some(KryoLazyStatsUtils.reduceFeatures(sft, hints)(_))
         BatchScanPlan(filter, attrTable, ranges, iters, Seq.empty, kvsToFeatures, reduce, attrThreads, hasDuplicates = false)
       } else {
         // check to see if we can execute against the index values
@@ -143,7 +143,7 @@ trait AttributeQueryableIndex extends AccumuloFeatureIndex with LazyLogging {
             filter.secondary.forall(IteratorTrigger.supportsFilter(indexSft, _))) {
           val iter = KryoLazyStatsIterator.configure(indexSft, this, filter.secondary, hints, hasDupes)
           val iters = visibilityIter(indexSft) :+ iter
-          val reduce = Some(KryoLazyStatsIterator.reduceFeatures(indexSft, hints)(_))
+          val reduce = Some(KryoLazyStatsUtils.reduceFeatures(indexSft, hints)(_))
           BatchScanPlan(filter, attrTable, ranges, iters, Seq.empty, kvsToFeatures, reduce, attrThreads, hasDuplicates = false)
         } else {
           // have to do a join against the record table
@@ -212,7 +212,7 @@ trait AttributeQueryableIndex extends AccumuloFeatureIndex with LazyLogging {
       // TODO GEOMESA-822 we can use the aggregating iterator if the features are kryo encoded
       (BinAggregatingIterator.nonAggregatedKvsToFeatures(sft, recordIndex, hints, SerializationType.KRYO), None)
     } else if (hints.isStatsIteratorQuery) {
-      (KryoLazyStatsIterator.kvsToFeatures(sft), Some(KryoLazyStatsIterator.reduceFeatures(sft, hints)(_)))
+      (KryoLazyStatsIterator.kvsToFeatures(sft), Some(KryoLazyStatsUtils.reduceFeatures(sft, hints)(_)))
     } else {
       (recordIndex.entriesToFeatures(sft, hints.getReturnSft), None)
     }
