+++ geomesa-accumulo/geomesa-accumulo-datastore/src/main/scala/org/locationtech/geomesa/accumulo/process/stats/StatsIteratorProcess.scala
@@ -16,10 +16,10 @@ import org.geotools.feature.DefaultFeatureCollection
 import org.geotools.feature.visitor.{AbstractCalcResult, CalcResult, FeatureCalc}
 import org.geotools.process.factory.{DescribeParameter, DescribeProcess, DescribeResult}
 import org.geotools.util.NullProgressListener
-import org.locationtech.geomesa.accumulo.iterators.KryoLazyStatsIterator
 import org.locationtech.geomesa.features.{ScalaSimpleFeature, TransformSimpleFeature}
 import org.locationtech.geomesa.index.api.QueryPlanner
 import org.locationtech.geomesa.index.conf.QueryHints
+import org.locationtech.geomesa.index.utils.KryoLazyStatsUtils
 import org.locationtech.geomesa.utils.geotools.GeometryUtils
 import org.locationtech.geomesa.utils.stats.Stat
 import org.opengis.feature.Feature
@@ -73,8 +73,6 @@ class StatsIteratorProcess extends LazyLogging {
 
 class StatsVisitor(features: SimpleFeatureCollection, statString: String, encode: Boolean, properties: Array[String] = null)
     extends FeatureCalc with LazyLogging {
-
-  import scala.collection.JavaConversions._
   val origSft = features.getSchema
 
   lazy val (transforms, transformSFT) = QueryPlanner.buildTransformSFT(origSft, properties)
@@ -88,7 +86,7 @@ class StatsVisitor(features: SimpleFeatureCollection, statString: String, encode
 
   lazy val stat: Stat = Stat(statSft, statString)
 
-  val returnSft = KryoLazyStatsIterator.StatsSft
+  val returnSft = KryoLazyStatsUtils.StatsSft
   val manualVisitResults: DefaultFeatureCollection = new DefaultFeatureCollection(null, returnSft)
   var resultCalc: StatsIteratorResult = null
 
@@ -108,7 +106,8 @@ class StatsVisitor(features: SimpleFeatureCollection, statString: String, encode
   override def getResult: CalcResult = {
     if (resultCalc == null) {
       val stats = if (encode) {
-        KryoLazyStatsIterator.encodeStat(stat, statSft)
+
+        KryoLazyStatsUtils.encodeStat(stat, statSft)
       } else {
         stat.toJson
       }
